requirements.txt
selenium==4.17.2
beautifulsoup4==4.12.3
pandas==2.2.0
lxml==5.1.0
pyarrow==15.0.0

langchain==0.1.5
langchain-openai==0.0.5
python-dotenv==1.0.1
faiss-cpu==1.7.4
# uncomment to use huggingface llms
# huggingface-hub

# uncomment to use instructor embeddings
#InstructorEmbedding==1.0.1
#sentence-transformers==2.2.2
fastapi==0.109.0
uvicorn==0.27.0.post1
SQLAlchemy==2.0.25
psycopg2==2.9.9
psycopg2-binary==2.9.9